Move code out of lib/solargraph_rails.rb, start some new abstractions
