+++ lib/solargraph_rails/files_loader.rb
@@ -1,0 +1,13 @@
+module SolargraphRails
+  class FilesLoader
+    def initialize(file_names)
+      @file_names = file_names
+    end
+
+    def each(&blk)
+      @file_names.each do |file_name|
+        blk.call(file_name, File.read(file_name))
+      end
+    end
+  end
+end
